requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
colorama==0.3.9
coverage==4.5.1
Flask==1.0.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.4
pycparser==2.18
PyJWT==1.6.4
pytest==3.6.3
pytest-cov==2.5.1
six==1.11.0
Werkzeug==0.14.1
